mitosheet: cleanup
